requirements.txt
jinja2
pyyaml
dnspython